+++ front/app/containers/Admin/pages/routes.ts
@@ -15,22 +15,4 @@ export default () => ({
       delay: 500,
     }),
   },
-  childRoutes: [
-    // {
-    //   path: 'new',
-    //   component: Loadable({
-    //     loader: () => import('./NewPageForm'),
-    //     loading: LoadableLoadingAdmin,
-    //     delay: 500,
-    //   }),
-    // },
-    {
-      path: ':pageId',
-      component: Loadable({
-        loader: () => import('./EditPageForm'),
-        loading: LoadableLoadingAdmin,
-        delay: 500,
-      }),
-    },
-  ],
 });
